backend/requirements.txt
pydantic==2.10.6
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
pandas==2.2.3
ta==0.11.0
python-binance==1.0.35
typer==0.15.2
python-dotenv==1.0.1
rich==13.9.4
yfinance==0.2.54
scikit-learn==1.6.1
requests==2.32.3
ccxt==4.4.61
numpy==1.26.4
pandas-ta==0.3.14b
altgraph==0.17.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
greenlet==3.1.1
idna==3.10
pefile==2023.2.7
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
urllib3==2.3.0
